refactor: [#31] extract modules and rename functions
