+++ tests/test_nautilus_librarian/test_typer/test_commands/test_workflows/test_gold_drawings_processing.py
